% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{fix_excel_dates}
\alias{fix_excel_dates}
\title{Fix Excel dates}
\usage{
fix_excel_dates(date)
}
\arguments{
\item{date}{date to be converted}
}
\description{
Add four years and a day to a given date if R has improperly read in dates
from an Excel file using the 1904 date system.
}
\examples{
fix_excel_dates("2014-05-01")
}
\references{
\url{http://datapub.cdlib.org/2014/04/10/abandon-all-hope-ye-who-enter-dates-in-excel/}
}

