% Generated by roxygen2 (4.0.2.9000): do not edit by hand
% Please edit documentation in R/read_all_files.R
\name{read_all_files}
\alias{read_all_files}
\title{Read all files}
\usage{
read_all_files(dir = ".", ext = "\\\\.(csv|CSV)$", FUN = read.csv, ...)
}
\arguments{
\item{dir}{directory where files are located. Defaults to the working directory.}

\item{ext}{file extension to be used. Defaults to '\\.(csv|CSV)$'.}

\item{FUN}{function to be used. Defaults to read.csv.}

\item{...}{optional arguments to FUN}
}
\description{
Read all files with a given file extension into a list. List elements are given names that correspond to the file name.
}

