% Generated by roxygen2 (4.0.2.9000): do not edit by hand
% Please edit documentation in R/title_case.R
\name{title_case}
\alias{title_case}
\title{Convert strings to Title Case (with spaces)}
\usage{
title_case(x)
}
\arguments{
\item{x}{string or vector of strings to convert to title case}
}
\description{
Convert strings to Title Case (with spaces)
}
\examples{
title_case("this is a sentence")
}
\author{
Kara Woo
}
\references{
\url{http://stackoverflow.com/a/11672354/2921990}
}

